elk/obs-monitor-api/requirements.txt
simpleobsws
elasticsearch[async]
uvicorn
fastapi
pyyaml
